+++ src/Calculators/Electronics/Basic/OhmsLaw/OhmsLaw.cs
@@ -34,6 +34,8 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
             "Ohm's Law",
             "The hammer in any electrical engineers toolbox. Calculate voltage, resistance and current using Ohm's law.",
             "pack://application:,,,/Calculators/Electronics/Basic/OhmsLaw/icon.png",
+            new string[]{ "Electronics", "Basic" },
+            new string[]{"ohm, resistor, resistance, voltage, current, law, vir"},
             new OhmsLawView()) {
 
             var view = (OhmsLawView)this.View;
@@ -46,8 +48,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                     "voltage",
                     view.TextBoxVoltageValue,
                     view.VoltageUnits,
-                    view.RadioButtonVoltage,
-                    //this.CalcVars,
+                    view.RadioButtonVoltage,                    
                     () => {
                         var current = this.Current.RawVal;
                         var resistance = this.Resistance.RawVal;
@@ -58,6 +59,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                         new NumberUnit("V", 1e0, NumberPreference.DEFAULT),
                         new NumberUnit("kV", 1e3),
                     },
+                    4,
                     Directions.Input,
                     null);
 
@@ -90,6 +92,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                         new NumberUnit("mA", 1e-3),
                         new NumberUnit("A", 1e0, NumberPreference.DEFAULT),
                     },
+                    4,
                     Directions.Input,
                     null);
 
@@ -107,8 +110,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                     "resistance",
                     view.TextBoxResistanceValue,
                     view.ResistanceUnits,
-                    view.RadioButtonResistance,
-                    //this.CalcVars,
+                    view.RadioButtonResistance,                    
                     () => {
                         var voltage = this.Voltage.RawVal;
                         var current = this.Current.RawVal;
@@ -121,6 +123,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.OhmsLaw
                         new NumberUnit("MΩ", 1e6),
                         new NumberUnit("GΩ", 1e9),
                     },
+                    4,
                     Directions.Output,
                     null);