+++ src/Calculators/Electronics/Basic/ResistorDivider/ResistorDividerCalc.cs
@@ -47,6 +47,8 @@ namespace NinjaCalc.Calculators.Electronics.Basic.ResistorDivider {
             "Resistor Divider",
             "Resistor dividers are a simple, widely-used circuit primitive for reducing a voltage based on a fixed ratio.",
             "pack://application:,,,/Calculators/Electronics/Basic/ResistorDivider/grid-icon.png",
+            new string[] { "Electronics", "Basic" },
+            new string[] { "resistor, resistance, voltage, divider, reduce" },
             new ResistorDividerCalcView()) {
 
             // Re-cast the view so we can access it's unique properties
@@ -73,6 +75,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.ResistorDivider {
                         new NumberUnit("V", 1e0, NumberPreference.DEFAULT),
                         new NumberUnit("kV", 1e3),
                     },
+                    4,
                     Directions.Input,
                     null);
 
@@ -105,6 +108,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.ResistorDivider {
                         new NumberUnit("MΩ", 1e6),
                         new NumberUnit("GΩ", 1e9),
                     },
+                    4,
                     Directions.Input,
                     null);
 
@@ -137,6 +141,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.ResistorDivider {
                         new NumberUnit("MΩ", 1e6),
                         new NumberUnit("GΩ", 1e9),
                     },
+                    4,
                     Directions.Input,
                     null);
 
@@ -167,6 +172,7 @@ namespace NinjaCalc.Calculators.Electronics.Basic.ResistorDivider {
                         new NumberUnit("V", 1e0, NumberPreference.DEFAULT),
                         new NumberUnit("kV", 1e3),
                     },
+                    4,
                     Directions.Output,
                     null);
 
@@ -197,7 +203,8 @@ namespace NinjaCalc.Calculators.Electronics.Basic.ResistorDivider {
                         new NumberUnit("uA", 1e-6),
                         new NumberUnit("mA", 1e-3, NumberPreference.DEFAULT),
                         new NumberUnit("A", 1e0),
-                });
+                },
+                4);
 
             // Add validators
             this.Iq.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));