+++ src/Calculators/Electronics/Pcb/TrackCurrentIpc2221A/TrackCurrentIpc2221ACalculator.cs
@@ -12,7 +12,7 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
 
     class TrackCurrentIpc2221ACalculator : Calculator {
 
-        CalcVarNumericalInput TraceCurrent {
+        CalcVarNumericalInput TrackCurrent {
             get;
             set;
         }
@@ -46,6 +46,8 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
             "Track Current (IPC-2221A)",
             "PCB track current carrying capability calculator, using the IPC-2221A standard.",
             "pack://application:,,,/Calculators/Electronics/Pcb/TrackCurrentIpc2221A/grid-icon.png",
+            new string[] { "Electronics", "PCB" },
+            new string[] { "pcb, track, current, trace, width, carry, heat, temperature, ipc, ipc2221a, ipc-2221a" },
             new TrackCurrentIpc2221AView()) {
 
             // Re-cast the view so we can access it's unique properties
@@ -55,7 +57,7 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
             //========================================= TRACE CURRENT =======================================//
             //===============================================================================================//
             
-           this.TraceCurrent = new CalcVarNumericalInput(
+           this.TrackCurrent = new CalcVarNumericalInput(
                 "traceCurrent",
                 view.TrackCurrentValue,
                 view.TrackCurrentUnits,                                                  
@@ -63,18 +65,19 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
                     new NumberUnit("mA", 1e-3),
                     new NumberUnit("A", 1e0, NumberPreference.DEFAULT),
                 },
+                4,
                 null);
 
             //===== VALIDATORS =====//
-            this.TraceCurrent.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
-            this.TraceCurrent.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
-            this.TraceCurrent.AddValidator(
+            this.TrackCurrent.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));
+            this.TrackCurrent.AddValidator(Validator.IsGreaterThanZero(CalcValidationLevels.Error));
+            this.TrackCurrent.AddValidator(
                 new Validator(() => {
-                    return ((this.TraceCurrent.RawVal > 35.0) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);                                      
+                    return ((this.TrackCurrent.RawVal > 35.0) ? CalcValidationLevels.Warning : CalcValidationLevels.Ok);                                      
                 },
                 "Current is above recommended maximum (35A). Equation will not be as accurate (extrapolation will occur)."));
 
-            this.CalcVars.Add(this.TraceCurrent);
+            this.CalcVars.Add(this.TrackCurrent);
 
             //===============================================================================================//
             //========================================== TEMP RISE ==========================================//
@@ -87,6 +90,7 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
                 new NumberUnit[]{
                     new NumberUnit("C", 1e0, NumberPreference.DEFAULT),                        
                 },
+                4,
                 null);
 
             //===== VALIDATORS =====//
@@ -117,6 +121,7 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
                     new NumberUnit("um", 1e-6, NumberPreference.DEFAULT),                        
                     new NumberUnit("mm", 1e-3),                        
                 },
+                4,
                 null);
 
             //===== VALIDATORS =====//
@@ -159,7 +164,7 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
                 view.MinTrackWidthUnits,
                 () => {
                     //Console.WriteLine("Equation() called for MinTrackWidth.");
-                    var traceCurrent = this.TraceCurrent.RawVal;
+                    var traceCurrent = this.TrackCurrent.RawVal;
                     var tempRise = this.TempRise.RawVal;
                     var trackThickness = this.TrackThickness.RawVal;
                     var trackLayer = this.TrackLayer.RawVal;
@@ -188,7 +193,8 @@ namespace NinjaCalc.Calculators.Electronics.Pcb.TrackCurrentIpc2221A {
                 new NumberUnit[]{
                     new NumberUnit("um", 1e-6),                        
                     new NumberUnit("mm", 1e-3, NumberPreference.DEFAULT),                        
-                });
+                },
+                4);
 
             // Add validators
             this.MinTrackWidth.AddValidator(Validator.IsNumber(CalcValidationLevels.Error));